Take out NodeState builder from test helper.
