+++ src/signer_node/mod.rs
@@ -3,10 +3,13 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 mod message_processor;
+pub mod node_state;
 mod node_parameters;
 mod utils;
 
-use crate::blockdata::Block;
+pub use node_parameters::NodeParameters;
+pub use crate::signer_node::node_state::NodeState;
+
 use crate::crypto::multi_party_schnorr::*;
 use crate::net::MessageType::{BlockGenerationRoundMessages, KeyGenerationMessage};
 use crate::net::{
@@ -25,7 +28,6 @@ use crate::util::*;
 use curv::cryptographic_primitives::secret_sharing::feldman_vss::VerifiableSS;
 use curv::elliptic::curves::traits::*;
 use curv::{FE, GE};
-pub use node_parameters::NodeParameters;
 use redis::ControlFlow;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
@@ -106,36 +108,6 @@ impl ToShares for SharedSecretMap {
 }
 
 static INITIAL_MASTER_INDEX: usize = 0;
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum NodeState {
-    Joining,
-    Master {
-        /// *block_key* is random value for using int the Signature Issuing Protocol.
-        /// VSS which is distributed to each other signer is generated by this key. All signers in
-        /// all block generation rounds has each own block_key.
-        block_key: Option<FE>,
-        shared_block_secrets: BidirectionalSharedSecretMap,
-        block_shared_keys: Option<(bool, FE, GE)>,
-        candidate_block: Option<Block>,
-        signatures: BTreeMap<SignerID, (FE, FE)>,
-        round_is_done: bool,
-    },
-    Member {
-        /// *block_key* is random value for using int the Signature Issuing Protocol.
-        /// VSS which is distributed to each other signer is generated by this key. All signers in
-        /// all block generation rounds has each own block_key.
-        block_key: Option<FE>,
-        shared_block_secrets: BidirectionalSharedSecretMap,
-        block_shared_keys: Option<(bool, FE, GE)>,
-        candidate_block: Option<Block>,
-        master_index: usize,
-    },
-    RoundComplete {
-        master_index: usize,
-        next_master_index: usize,
-    },
-}
 
 impl<T: TapyrusApi, C: ConnectionManager> SignerNode<T, C> {
     pub fn new(connection_manager: C, params: NodeParameters<T>) -> Self
